fix: xcode 11, iOS 13 @main error
